@@ -3,19 +3,19 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "axios": "^0.19.2",
-    "leaflet": "^1.6.0",
-    "moment": "^2.26.0",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
-    "react-leaflet": "^2.7.0",
-    "react-moment": "^0.9.7",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "3.0.1",
-    "redux": "^4.0.5",
+    "axios": "^1.4.0",
+    "leaflet": "^1.9.4",
+    "moment": "^2.29.4",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-leaflet": "^4.2.1",
+    "react-moment": "^1.1.3",
+    "react-router-dom": "^6.14.1",
+    "react-scripts": "5.0.1",
+    "redux": "^4.2.1",
     "redux-devtools-extension": "^2.13.8",
-    "redux-thunk": "^2.3.0",
-    "uuid": "^3.4.0"
+    "redux-thunk": "^2.4.2",
+    "uuid": "^9.0.0"
   },
   "scripts": {
     "start": "react-scripts start",
