@@ -8,20 +8,28 @@ TODO:
 
 import React, { useEffect } from 'react';
 import './App.css';
-import { BrowserRouter as Router, Route, Switch } from 'react-router-dom';
 
 // Components
 import Navigation from './components/layout/Navigation';
+import Alert from './components/layout/Alert';
+import Login from './components/auth/Login';
 import Breweries from './components/Breweries';
 import Contact from './components/Contact';
 import About from './components/About';
 import Brewery from './components/brewery/Brewery';
-import Routes from './components/routing/Routes';
+import CreateBrewery from './components/brewery-forms/CreateBrewery';
+import EditBrewery from './components/brewery-forms/EditBrewery';
+import AddBeer from './components/brewery-forms/AddBeer';
+import NotFound from './components/layout/NotFound';
 import Footer from './components/layout/Footer';
+
+import PrivateRoute from './components/routing/PrivateRoute';
 
 import store from './store';
 import { loadUser } from './actions/auth';
 import setAuthToken from './utils/setAuthToken';
+
+import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
 
 function App() {
   useEffect(() => {
@@ -35,13 +43,18 @@ function App() {
     <Router>
       <div className="App">
         <Navigation />
-          <Switch>
+        <Alert />
+          <Routes>
             <Route exact path='/' component={Breweries} />
             <Route exact path='/about' component={About} />
             <Route exact path='/contact' component={Contact} />
             <Route exact path='/breweries/:id' component={Brewery} />
-            <Route component={Routes} />
-          </Switch>
+            <Route exact path='/login' component={Login} />
+            <PrivateRoute exact path='/new-brewery' component={CreateBrewery} />
+            <PrivateRoute exact path='/breweries/:id/edit' component={EditBrewery} />
+            <PrivateRoute exact path='/breweries/:id/beers' component={AddBeer} />
+            <Route component={NotFound} />
+          </Routes>
         <Footer />
       </div>
     </Router>
