@@ -1,9 +1,9 @@
 // Feedback for post submissions and auth routes
 import { SET_ALERT, REMOVE_ALERT } from './types';
-import uuid from 'uuid';
+import { v4 as uuidv4 } from 'uuid';
 
 export const setAlert = (msg, alertType, timeout=5000) => dispatch => {
-  const id = uuid.v4();
+  const id = uuidv4();
   dispatch({
     type: SET_ALERT,
     payload: {
