@@ -2,7 +2,7 @@ import { SET_ALERT, REMOVE_ALERT } from '../actions/types';
 
 const initialState = [];
 
-export default function auth(state = initialState, action) {
+export default function authReducer(state = initialState, action) {
   const { type, payload } = action;
   switch (type) {
     case SET_ALERT:
