@@ -14,7 +14,7 @@ const initialState = {
   loading: true
 };
 
-export default function(state = initialState, action) {
+export default function authReducer(state = initialState, action) {
   const { type, payload } = action;
   switch (type) {
     case USER_LOADED:
