@@ -18,7 +18,7 @@ const initialState = {
   loading: true
 };
 
-export default function(state = initialState, action) {
+export default function breweryReducer(state = initialState, action) {
   const { type, payload } = action;
 
   switch (type) {
